change add button to special activity
just like the "Choose weekly activities" special activity.

I opt out of the "+" button because it is too prominent. New
users will think it is the main add action in the app.

Also, having it as an activity reminds you to decide your one
time tasks to do each day, and marks itself as done when you do
so, allowing you to feel good about the one time tasks you done
and not thinking "I can do more!".