add framework for the new activity handling
The intention behind this changes: I want all code about one
activity type centralized and independent from other activity
types. This way we'll only have to change code in one class when
we want to modify an activity type, or add a new class when we
want to create a new one.