cretate NewActivitySlice
this new version of the activitySlice implements the functionality of
both the activitySlice and the activityRecordsSlice. I aggregate the two
into one since the data of both can be more easily managed and accesed
in a single slice.

It also includes the improvements needed for the completion-rates
feature.

This slice is not in use yet.