make repetitions a default field of entries
It did not make much sense to have it as "extra data".
Not every activity type uses the repetitions field, but managing things
like activity type changes is more complex if we treat the repetitions
as an optional field.

I've added a function in the activity handler called "usesRepetitions"
so that each activity type can choose to return true so that the
repetitions field appears in its activitydetailscreen.