fix goals and activities day regresion
Before this commit, if you edited a goal or an activity and then
returned to the previous day by changing the dayStartHour,
the changes in the goal and activity were lost, and recovered
when you returned to the next day.

With this commit the chenges you made are moved, as if you did them on
the previous day.