fix StatsScreen dropdown picker
The dropdown picker in the StatsScreen rendered under other components
if used three times in a row, making the options not clickable.
Fixed it by making the dropdown relatively positioned
instead
of absolutelly. Now it displaces components under it (in the Y axis)
instead of avobe them (in the Z axis).