fix restored goal list item not responding
Just after restoring a goal, its list item did not respond to touch
in the GoalsScreen. After relaunching the app, the item was responding again.

This is likely due to an issue with FlatList, as they have arisen in the past.
I am just changing the flat list by a map call for a quick fix.