fix getWeeklyStats bug
In getWeeklyStats we recieve a adjusted date (meaning that it already
has into account the day start hour). But we were adjusting it once
more. Adjusting it twice means that in some cases we get the staties of
the day before the intended date.

This commit removes the second adjust.